Added Check for Goal and Double Tile
- Added a check to see if you are at the goal before "killing the player"

- Added a "2 use" tile that lets you move over it 2 times before being destroyed

- Setup part of the GameManager script for future use

Side notes: The games core controls and mechanics are being created smoothly and quickly! I'm really surprised how much I've been doing and am excited to keep working on it and hopefully figure out a place and way to publish it. I have a lot more "things" I want to add that will really take it to the next level, but that will come after I finish all of the main basic stuff.