Added Title Screen + More
- Added the start of a temporary title screen

- Setup more core mechanics of the game

Side notes: So I was planning to add a sort of "multitask" element to the game but I ultimatly decided to scrap the idea and maybe save it for a future game. After starting to build it into the game, it just didn't feel like it fit well with the main game, plus it would eat up a lot of time that I could be spending on making the core game good.

I'll save the idea in case I decide to use it in a future game and as for this game, I am going to focus on a simpler idea to get it completed and polished.