Major code changes
- Lots of code changes (moving code around, recoding things, deleting scripts, cleaning up scripts, and more)

Side notes: So, without going into too much detail, this update was focused on MAJOR code cleanup and reworks. After starting this project and learning a lot more about coding and Unity in general, a lot of the project just felt messy to me. It bothered me too much to ignore so I took the time to go through pretty much every script and figure out a way to clean a lot of it up and scrap what I didn't need.

A huge change that I made was deciding to separate the levels into different scenes. As of now, there is only the first level in the second scene, but after some more work I should be able to have a decent Level Template that I can make a prefab out of to create more levels.

The next few things on my list is: what I just described, creating a death screen, and creating scripts for the death screen and level complete screen to properly function.

After I get those done (hopefully by tomorrow or the day after) I can start working on some other parts of the project like more levels and mechanics.