Load system setup + First test build
- Setup system to continue the game where you left off

- Setup continue and new game buttons in the main menu

- Added a second level

- Added tips text to display before starting a level to help the player learn how to play

- Added the first test build of the game!

Dev notes: So I finished the system to continue where you left off which is pretty cool! Also I made a build folder with the first build of the game. It's not much since its only two short levels, but I wanted to make sure everything is working as it should and it is!

As long as nothing major comes up, I should be able to start working on level designs and see if I decide to add more features to make the game more interesting. I am confident everything should be setup to move on to this next step.

As I'm making more of the game I'll be thinking up ideas of the name of the game, art ideas, and looking for sounds and music for the game.

Sounds and music will most likely be the last thing I do (Other than testing and polishing), with art being right before it. I'm excited to continue development of this "soon to be named" game! Happy holidays everyone!